message cell is completed

+++ Flash Chat iOS13/Model/Message.swift
@@ -11,5 +11,5 @@ import Foundation
 struct Message {
     let from: String
     let message: String
-    let sentDate: Int64
+    let sentDate: Double
 }
